+++ src/main/java/cn/nukkit/Player.java
@@ -4191,9 +4191,15 @@ public class Player extends EntityHuman implements CommandSender, InventoryHolde
 
         this.sendExperienceLevel(level);
         this.sendExperience(exp);
-        if (playLevelUpSound && levelBefore / 5 != level / 5 && this.lastPlayerdLevelUpSoundTime < this.age - 100) {
+        if (playLevelUpSound && levelBefore < level && levelBefore / 5 != level / 5 && this.lastPlayerdLevelUpSoundTime < this.age - 100) {
             this.lastPlayerdLevelUpSoundTime = this.age;
-            this.level.addSound(this, Sound.RANDOM_LEVELUP, 1F, 1F, this);
+            this.level.addLevelSoundEvent(
+                    this, 
+                    LevelSoundEventPacketV2.SOUND_LEVELUP,
+                    Math.min(7, level / 5) << 28,
+                    "",
+                    false, false
+            );
         }
     }